backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.1.0
setuptools
django-allauth==0.57.0
openai==1.3.0
google-generativeai==0.3.2
anthropic==0.18.1
requests==2.31.0

